improved dht server handling, some refactoring

+++ src/main/java/org/ton/ui/controllers/MainController.java
@@ -1449,11 +1449,6 @@ public class MainController implements Initializable {
                 validationTabs.getTabs().add(getNodeTabByName(n));
             }
         }
-//        for (String nodeName : settings.getActiveNodes()) {
-//            if (!nodeName.contains("genesis")) {
-//                validationTabs.getTabs().add(getNodeTabByName(nodeName));
-//            }
-//        }
     }
 
     public Tab getNodeTabByName(String nodeName) {
@@ -2420,45 +2415,50 @@ public class MainController implements Initializable {
 
         newNodeExecutorService.execute(() -> {
             Thread.currentThread().setName("MyLocalTon - Creating validator");
-            try {
-                mainController.addValidatorBtn.setDisable(true);
+            if (Long.parseLong(mainController.currentBlockNum.getText()) > 120) {
+                try {
+                    mainController.addValidatorBtn.setDisable(true);
 
-                org.ton.settings.Node node = Utils.getNewNode();
-                if (nonNull(node)) {
-                    log.info("creating validator {}", node.getNodeName());
-                    App.mainController.showInfoMsg("Creating validator " + node.getNodeName() + ". You will be informed once it's finished.", 8);
+                    org.ton.settings.Node node = Utils.getNewNode();
+                    if (nonNull(node)) {
+                        log.info("creating validator {}", node.getNodeName());
+                        App.mainController.showInfoMsg("Creating validator " + node.getNodeName() + ". You will be informed once it's finished.", 8);
+                        Thread.sleep(500);
 
-                    //delete unfinished node creation
-                    FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + node.getNodeName()));
+                        //delete unfinished node creation
+                        FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + node.getNodeName()));
 
-                    MyLocalTon.getInstance().createFullnode(node, true, true);
+                        MyLocalTon.getInstance().createFullnode(node, true, true);
 
-                    if (isWindows()) {
-                        Utils.waitForBlockchainReady(node);
-                        Utils.waitForNodeSynchronized(node);
+                        if (isWindows()) {
+                            Utils.waitForBlockchainReady(node);
+                            Utils.waitForNodeSynchronized(node);
+                        }
+
+                        Tab newTab = Utils.getNewNodeTab();
+                        Platform.runLater(() -> {
+                            validationTabs.getTabs().add(newTab);
+                        });
+
+                        settings.getActiveNodes().add(node.getNodeName());
+                        MyLocalTon.getInstance().saveSettingsToGson();
+                        mainController.addValidatorBtn.setDisable(false);
+
+                        // FYI. Status of all nodes reported back from the thread "Node Monitor" and shown on a corresponding tab
+
+                        //App.mainController.showInfoMsg("Validator " + node.getNodeName() + " has been successfully created", 5);
+                        showDialogMessage("Completed", "Validator " + node.getNodeName() + " has been successfully created, now synchronizing. Once elections will be opened it will take part in them.");
+                    } else {
+                        showDialogMessage("The limit has been reached", "It is possible to have up to 6 additional validators. The first one is reserved, thus in total you may have 7 validators.");
                     }
-
-                    Tab newTab = Utils.getNewNodeTab();
-                    Platform.runLater(() -> {
-                        validationTabs.getTabs().add(newTab);
-                    });
-
-                    settings.getActiveNodes().add(node.getNodeName());
-                    MyLocalTon.getInstance().saveSettingsToGson();
+                } catch (Exception e) {
+                    log.error("Error creating validator: {}", e.getMessage());
+                    App.mainController.showErrorMsg("Error creating validator", 3);
+                } finally {
                     mainController.addValidatorBtn.setDisable(false);
-
-                    // FYI. Status of all nodes reported back from the thread "Node Monitor" and shown on a corresponding tab
-
-                    //App.mainController.showInfoMsg("Validator " + node.getNodeName() + " has been successfully created", 5);
-                    showDialogMessage("Completed", "Validator " + node.getNodeName() + " has been successfully created, now synchronizing. Once elections will be opened it will take part in them.");
-                } else {
-                    showDialogMessage("The limit has been reached", "It is possible to have up to 6 additional validators. The first one is reserved, thus in total you may have 7 validators.");
                 }
-            } catch (Exception e) {
-                log.error("Error creating validator: {}", e.getMessage());
-                App.mainController.showErrorMsg("Error creating validator", 3);
-            } finally {
-                mainController.addValidatorBtn.setDisable(false);
+            } else {
+                showDialogMessage("Too early", "Please wait for 120 blocks to be generated. At least one validator group should be rotated.");
             }
         });
         newNodeExecutorService.shutdown();
